--- main.py
@@ -1,3 +1,0 @@
-print("hello, Doctor AI!")
-
-
